requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.40.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
wsproto==1.2.0
